separated models into smaller files
